fixes django tests
